restructured added gitignore
